Replace fuzzy matching algorithm with improved fzy-based implementation
This version is based on the original fzy algorithm, with one key enhancement:
multibyte character support.

The original implementation supports only ASCII via table lookups for
uppercase/lowercase scoring.

This patch replaces those lookups with function calls, making it compatible
with multibyte character sets.

Core logic (match_row() and match_positions()) remains faithful to the
original, but now operates on codepoints rather than single-byte characters.
