+++ src/proto.h
@@ -188,6 +188,7 @@ void mbyte_im_set_active(int active_arg);
 # if defined(FEAT_CRYPT) || defined(FEAT_PERSISTENT_UNDO)
 #  include "sha256.pro"
 # endif
+# include "fuzzy.pro"
 # include "search.pro"
 # ifdef FEAT_SIGNS
 #  include "sign.pro"
