+++ src/search.c
@@ -42,17 +42,6 @@ static void find_mps_values(int *initc, int *findc, int *backwards, int switchit
 static int is_zero_width(char_u *pattern, size_t patternlen, int move, pos_T *cur, int direction);
 static void cmdline_search_stat(int dirc, pos_T *pos, pos_T *cursor_pos, int show_top_bot_msg, char_u *msgbuf, size_t msgbuflen, int recompute, int maxcount, long timeout);
 static void update_search_stat(int dirc, pos_T *pos, pos_T *cursor_pos, searchstat_T *stat, int recompute, int maxcount, long timeout);
-static int fuzzy_match_compute_score(char_u *fuzpat, char_u *str, int strSz, int_u *matches, int numMatches, int camelcase);
-static int fuzzy_match_recursive(char_u *fuzpat, char_u *str, int_u strIdx, int *outScore, char_u *strBegin, int strLen, int_u *srcMatches, int_u *matches, int maxMatches, int nextMatch, int *recursionCount, int camelcase);
-#if defined(FEAT_EVAL) || defined(FEAT_PROTO)
-static int fuzzy_match_item_compare(const void *s1, const void *s2);
-static void fuzzy_match_in_list(list_T *l, char_u *str, int matchseq, char_u *key, callback_T *item_cb, int retmatchpos, list_T *fmatchlist, long max_matches, int camelcase);
-static void do_fuzzymatch(typval_T *argvars, typval_T *rettv, int retmatchpos);
-#endif
-static int fuzzy_match_str_compare(const void *s1, const void *s2);
-static void fuzzy_match_str_sort(fuzmatch_str_T *fm, int sz);
-static int fuzzy_match_func_compare(const void *s1, const void *s2);
-static void fuzzy_match_func_sort(fuzmatch_str_T *fm, int sz);
 
 #define SEARCH_STAT_DEF_TIMEOUT 40L
 // 'W ':  2 +
@@ -4290,1190 +4279,3 @@ the_end:
 #endif
 }
 #endif
-
-/*
- * Fuzzy string matching
- *
- * Ported from the lib_fts library authored by Forrest Smith.
- * https://github.com/forrestthewoods/lib_fts/tree/master/code
- *
- * The following blog describes the fuzzy matching algorithm:
- * https://www.forrestthewoods.com/blog/reverse_engineering_sublime_texts_fuzzy_match/
- *
- * Each matching string is assigned a score. The following factors are checked:
- *   - Matched letter
- *   - Unmatched letter
- *   - Consecutively matched letters
- *   - Proximity to start
- *   - Letter following a separator (space, underscore)
- *   - Uppercase letter following lowercase (aka CamelCase)
- *
- * Matched letters are good. Unmatched letters are bad. Matching near the start
- * is good. Matching the first letter in the middle of a phrase is good.
- * Matching the uppercase letters in camel case entries is good.
- *
- * The score assigned for each factor is explained below.
- * File paths are different from file names. File extensions may be ignorable.
- * Single words care about consecutive matches but not separators or camel
- * case.
- *   Score starts at 100
- *   Matched letter: +0 points
- *   Unmatched letter: -1 point
- *   Consecutive match bonus: +15 points
- *   First letter bonus: +15 points
- *   Separator bonus: +30 points
- *   Camel case bonus: +30 points
- *   Unmatched leading letter: -5 points (max: -15)
- *
- * There is some nuance to this. Scores don’t have an intrinsic meaning. The
- * score range isn’t 0 to 100. It’s roughly [50, 150]. Longer words have a
- * lower minimum score due to unmatched letter penalty. Longer search patterns
- * have a higher maximum score due to match bonuses.
- *
- * Separator and camel case bonus is worth a LOT. Consecutive matches are worth
- * quite a bit.
- *
- * There is a penalty if you DON’T match the first three letters. Which
- * effectively rewards matching near the start. However there’s no difference
- * in matching between the middle and end.
- *
- * There is not an explicit bonus for an exact match. Unmatched letters receive
- * a penalty. So shorter strings and closer matches are worth more.
- */
-typedef struct
-{
-    int		idx;		// used for stable sort
-    listitem_T	*item;
-    int		score;
-    list_T	*lmatchpos;
-} fuzzyItem_T;
-
-// bonus for adjacent matches; this is higher than SEPARATOR_BONUS so that
-// matching a whole word is preferred.
-#define SEQUENTIAL_BONUS 40
-// bonus if match occurs after a path separator
-#define PATH_SEPARATOR_BONUS 30
-// bonus if match occurs after a word separator
-#define WORD_SEPARATOR_BONUS 25
-// bonus if match is uppercase and prev is lower
-#define CAMEL_BONUS 30
-// bonus if the first letter is matched
-#define FIRST_LETTER_BONUS 15
-// bonus if exact match
-#define EXACT_MATCH_BONUS 100
-// bonus if case match when no ignorecase
-#define CASE_MATCH_BONUS 25
-// penalty applied for every letter in str before the first match
-#define LEADING_LETTER_PENALTY (-5)
-// maximum penalty for leading letters
-#define MAX_LEADING_LETTER_PENALTY (-15)
-// penalty for every letter that doesn't match
-#define UNMATCHED_LETTER_PENALTY (-1)
-// penalty for gap in matching positions (-2 * k)
-#define GAP_PENALTY	(-2)
-// Score for a string that doesn't fuzzy match the pattern
-#define SCORE_NONE	(-9999)
-
-#define FUZZY_MATCH_RECURSION_LIMIT	10
-
-/*
- * Compute a score for a fuzzy matched string. The matching character locations
- * are in 'matches'.
- */
-    static int
-fuzzy_match_compute_score(
-	char_u		*fuzpat,
-	char_u		*str,
-	int		strSz,
-	int_u		*matches,
-	int		numMatches,
-	int		camelcase)
-{
-    int		score;
-    int		penalty;
-    int		unmatched;
-    int		i;
-    char_u	*p = str;
-    int_u	sidx = 0;
-    int		is_exact_match = TRUE;
-    char_u	*orig_fuzpat = fuzpat - numMatches;
-    char_u	*curpat = orig_fuzpat;
-    int		pat_idx = 0;
-    // Track consecutive camel case matches
-    int		consecutive_camel = 0;
-
-    // Initialize score
-    score = 100;
-
-    // Apply leading letter penalty
-    penalty = LEADING_LETTER_PENALTY * matches[0];
-    if (penalty < MAX_LEADING_LETTER_PENALTY)
-	penalty = MAX_LEADING_LETTER_PENALTY;
-    score += penalty;
-
-    // Apply unmatched penalty
-    unmatched = strSz - numMatches;
-    score += UNMATCHED_LETTER_PENALTY * unmatched;
-    // In a long string, not all matches may be found due to the recursion limit.
-    // If at least one match is found, reset the score to a non-negative value.
-    if (score < 0 && numMatches > 0)
-	score = 0;
-
-    // Apply ordering bonuses
-    for (i = 0; i < numMatches; ++i)
-    {
-	int_u	currIdx = matches[i];
-	int	curr;
-	int	is_camel = FALSE;
-
-	if (i > 0)
-	{
-	    int_u	prevIdx = matches[i - 1];
-
-	    // Sequential
-	    if (currIdx == (prevIdx + 1))
-		score += SEQUENTIAL_BONUS;
-	    else
-	    {
-		score += GAP_PENALTY * (currIdx - prevIdx);
-		// Reset consecutive camel count on gap
-		consecutive_camel = 0;
-	    }
-	}
-
-	// Check for bonuses based on neighbor character value
-	if (currIdx > 0)
-	{
-	    // Camel case
-	    int neighbor = ' ';
-
-	    if (has_mbyte)
-	    {
-		while (sidx < currIdx)
-		{
-		    neighbor = (*mb_ptr2char)(p);
-		    MB_PTR_ADV(p);
-		    sidx++;
-		}
-		curr = (*mb_ptr2char)(p);
-	    }
-	    else
-	    {
-		neighbor = str[currIdx - 1];
-		curr = str[currIdx];
-	    }
-
-	    // Enhanced camel case scoring
-	    if (camelcase && vim_islower(neighbor) && vim_isupper(curr))
-	    {
-		score += CAMEL_BONUS * 2;  // Double the camel case bonus
-		is_camel = TRUE;
-		consecutive_camel++;
-		// Additional bonus for consecutive camel
-		if (consecutive_camel > 1)
-		    score += CAMEL_BONUS;
-	    }
-	    else
-		consecutive_camel = 0;
-
-	    // Bonus if the match follows a separator character
-	    if (neighbor == '/' || neighbor == '\\')
-		score += PATH_SEPARATOR_BONUS;
-	    else if (neighbor == ' ' || neighbor == '_')
-		score += WORD_SEPARATOR_BONUS;
-	}
-	else
-	{
-	    // First letter
-	    score += FIRST_LETTER_BONUS;
-	    curr = has_mbyte ? (*mb_ptr2char)(p) : str[currIdx];
-	}
-
-	// Case matching bonus
-	if (vim_isalpha(curr))
-	{
-	    while (pat_idx < i && *curpat)
-	    {
-		if (has_mbyte)
-		    MB_PTR_ADV(curpat);
-		else
-		    curpat++;
-		pat_idx++;
-	    }
-
-	    if (has_mbyte)
-	    {
-		if (curr == (*mb_ptr2char)(curpat))
-		{
-		    score += CASE_MATCH_BONUS;
-		    // Extra bonus for exact case match in camel
-		    if (is_camel)
-			score += CASE_MATCH_BONUS / 2;
-		}
-	    }
-	    else if (curr == *curpat)
-	    {
-		score += CASE_MATCH_BONUS;
-		if (is_camel)
-		    score += CASE_MATCH_BONUS / 2;
-	    }
-	}
-
-	// Check exact match condition
-	if (currIdx != (int_u)i)
-	    is_exact_match = FALSE;
-    }
-
-    // Boost score for exact matches
-    if (is_exact_match && numMatches == strSz)
-	score += EXACT_MATCH_BONUS;
-
-    return score;
-}
-
-/*
- * Perform a recursive search for fuzzy matching 'fuzpat' in 'str'.
- * Return the number of matching characters.
- */
-    static int
-fuzzy_match_recursive(
-	char_u		*fuzpat,
-	char_u		*str,
-	int_u		strIdx,
-	int		*outScore,
-	char_u		*strBegin,
-	int		strLen,
-	int_u		*srcMatches,
-	int_u		*matches,
-	int		maxMatches,
-	int		nextMatch,
-	int		*recursionCount,
-	int		camelcase)
-{
-    // Recursion params
-    int		recursiveMatch = FALSE;
-    int_u	bestRecursiveMatches[MAX_FUZZY_MATCHES];
-    int		bestRecursiveScore = 0;
-    int		first_match;
-    int		matched;
-
-    // Count recursions
-    ++*recursionCount;
-    if (*recursionCount >= FUZZY_MATCH_RECURSION_LIMIT)
-	return 0;
-
-    // Detect end of strings
-    if (*fuzpat == NUL || *str == NUL)
-	return 0;
-
-    // Loop through fuzpat and str looking for a match
-    first_match = TRUE;
-    while (*fuzpat != NUL && *str != NUL)
-    {
-	int	c1;
-	int	c2;
-
-	c1 = PTR2CHAR(fuzpat);
-	c2 = PTR2CHAR(str);
-
-	// Found match
-	if (vim_tolower(c1) == vim_tolower(c2))
-	{
-	    // Supplied matches buffer was too short
-	    if (nextMatch >= maxMatches)
-		return 0;
-
-	    int		recursiveScore = 0;
-	    int_u	recursiveMatches[MAX_FUZZY_MATCHES];
-	    CLEAR_FIELD(recursiveMatches);
-
-	    // "Copy-on-Write" srcMatches into matches
-	    if (first_match && srcMatches)
-	    {
-		memcpy(matches, srcMatches, nextMatch * sizeof(srcMatches[0]));
-		first_match = FALSE;
-	    }
-
-	    // Recursive call that "skips" this match
-	    char_u *next_char = str + (has_mbyte ? (*mb_ptr2len)(str) : 1);
-	    if (fuzzy_match_recursive(fuzpat, next_char, strIdx + 1,
-			&recursiveScore, strBegin, strLen, matches,
-			recursiveMatches,
-			ARRAY_LENGTH(recursiveMatches),
-			nextMatch, recursionCount, camelcase))
-	    {
-		// Pick best recursive score
-		if (!recursiveMatch || recursiveScore > bestRecursiveScore)
-		{
-		    memcpy(bestRecursiveMatches, recursiveMatches,
-			    MAX_FUZZY_MATCHES * sizeof(recursiveMatches[0]));
-		    bestRecursiveScore = recursiveScore;
-		}
-		recursiveMatch = TRUE;
-	    }
-
-	    // Advance
-	    matches[nextMatch++] = strIdx;
-	    if (has_mbyte)
-		MB_PTR_ADV(fuzpat);
-	    else
-		++fuzpat;
-	}
-	if (has_mbyte)
-	    MB_PTR_ADV(str);
-	else
-	    ++str;
-	strIdx++;
-    }
-
-    // Determine if full fuzpat was matched
-    matched = *fuzpat == NUL ? TRUE : FALSE;
-
-    // Calculate score
-    if (matched)
-	*outScore = fuzzy_match_compute_score(fuzpat, strBegin, strLen, matches,
-		nextMatch, camelcase);
-
-    // Return best result
-    if (recursiveMatch && (!matched || bestRecursiveScore > *outScore))
-    {
-	// Recursive score is better than "this"
-	memcpy(matches, bestRecursiveMatches, maxMatches * sizeof(matches[0]));
-	*outScore = bestRecursiveScore;
-	return nextMatch;
-    }
-    else if (matched)
-	return nextMatch;	// "this" score is better than recursive
-
-    return 0;		// no match
-}
-
-/*
- * fuzzy_match()
- *
- * Performs exhaustive search via recursion to find all possible matches and
- * match with highest score.
- * Scores values have no intrinsic meaning.  Possible score range is not
- * normalized and varies with pattern.
- * Recursion is limited internally (default=10) to prevent degenerate cases
- * (pat_arg="aaaaaa" str="aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa").
- * Uses char_u for match indices. Therefore patterns are limited to
- * MAX_FUZZY_MATCHES characters.
- *
- * Returns TRUE if "pat_arg" matches "str". Also returns the match score in
- * "outScore" and the matching character positions in "matches".
- */
-    int
-fuzzy_match(
-	char_u		*str,
-	char_u		*pat_arg,
-	int		matchseq,
-	int		*outScore,
-	int_u		*matches,
-	int		maxMatches,
-	int		camelcase)
-{
-    int		recursionCount = 0;
-    int		len = MB_CHARLEN(str);
-    char_u	*save_pat;
-    char_u	*pat;
-    char_u	*p;
-    int		complete = FALSE;
-    int		score = 0;
-    int		numMatches = 0;
-    int		matchCount;
-
-    *outScore = 0;
-
-    save_pat = vim_strsave(pat_arg);
-    if (save_pat == NULL)
-	return FALSE;
-    pat = save_pat;
-    p = pat;
-
-    // Try matching each word in 'pat_arg' in 'str'
-    while (TRUE)
-    {
-	if (matchseq)
-	    complete = TRUE;
-	else
-	{
-	    // Extract one word from the pattern (separated by space)
-	    p = skipwhite(p);
-	    if (*p == NUL)
-		break;
-	    pat = p;
-	    while (*p != NUL && !VIM_ISWHITE(PTR2CHAR(p)))
-	    {
-		if (has_mbyte)
-		    MB_PTR_ADV(p);
-		else
-		    ++p;
-	    }
-	    if (*p == NUL)		// processed all the words
-		complete = TRUE;
-	    *p = NUL;
-	}
-
-	score = 0;
-	recursionCount = 0;
-	matchCount = fuzzy_match_recursive(pat, str, 0, &score, str, len, NULL,
-				matches + numMatches, maxMatches - numMatches,
-				0, &recursionCount, camelcase);
-	if (matchCount == 0)
-	{
-	    numMatches = 0;
-	    break;
-	}
-
-	// Accumulate the match score and the number of matches
-	*outScore += score;
-	numMatches += matchCount;
-
-	if (complete)
-	    break;
-
-	// try matching the next word
-	++p;
-    }
-
-    vim_free(save_pat);
-    return numMatches != 0;
-}
-
-#if defined(FEAT_EVAL) || defined(FEAT_PROTO)
-/*
- * Sort the fuzzy matches in the descending order of the match score.
- * For items with same score, retain the order using the index (stable sort)
- */
-    static int
-fuzzy_match_item_compare(const void *s1, const void *s2)
-{
-    int		v1 = ((fuzzyItem_T *)s1)->score;
-    int		v2 = ((fuzzyItem_T *)s2)->score;
-    int		idx1 = ((fuzzyItem_T *)s1)->idx;
-    int		idx2 = ((fuzzyItem_T *)s2)->idx;
-
-    if (v1 == v2)
-	return idx1 == idx2 ? 0 : idx1 > idx2 ? 1 : -1;
-    else
-	return v1 > v2 ? -1 : 1;
-}
-
-/*
- * Fuzzy search the string 'str' in a list of 'items' and return the matching
- * strings in 'fmatchlist'.
- * If 'matchseq' is TRUE, then for multi-word search strings, match all the
- * words in sequence.
- * If 'items' is a list of strings, then search for 'str' in the list.
- * If 'items' is a list of dicts, then either use 'key' to lookup the string
- * for each item or use 'item_cb' Funcref function to get the string.
- * If 'retmatchpos' is TRUE, then return a list of positions where 'str'
- * matches for each item.
- */
-    static void
-fuzzy_match_in_list(
-	list_T		*l,
-	char_u		*str,
-	int		matchseq,
-	char_u		*key,
-	callback_T	*item_cb,
-	int		retmatchpos,
-	list_T		*fmatchlist,
-	long		max_matches,
-	int		camelcase)
-{
-    long	len;
-    fuzzyItem_T	*items;
-    listitem_T	*li;
-    long	i = 0;
-    long	match_count = 0;
-    int_u	matches[MAX_FUZZY_MATCHES];
-
-    len = list_len(l);
-    if (len == 0)
-	return;
-    if (max_matches > 0 && len > max_matches)
-	len = max_matches;
-
-    items = ALLOC_CLEAR_MULT(fuzzyItem_T, len);
-    if (items == NULL)
-	return;
-
-    // For all the string items in items, get the fuzzy matching score
-    FOR_ALL_LIST_ITEMS(l, li)
-    {
-	int		score;
-	char_u		*itemstr;
-	typval_T	rettv;
-
-	if (max_matches > 0 && match_count >= max_matches)
-	    break;
-
-	itemstr = NULL;
-	rettv.v_type = VAR_UNKNOWN;
-	if (li->li_tv.v_type == VAR_STRING)	// list of strings
-	    itemstr = li->li_tv.vval.v_string;
-	else if (li->li_tv.v_type == VAR_DICT
-				&& (key != NULL || item_cb->cb_name != NULL))
-	{
-	    // For a dict, either use the specified key to lookup the string or
-	    // use the specified callback function to get the string.
-	    if (key != NULL)
-		itemstr = dict_get_string(li->li_tv.vval.v_dict,
-							   (char *)key, FALSE);
-	    else
-	    {
-		typval_T	argv[2];
-
-		// Invoke the supplied callback (if any) to get the dict item
-		li->li_tv.vval.v_dict->dv_refcount++;
-		argv[0].v_type = VAR_DICT;
-		argv[0].vval.v_dict = li->li_tv.vval.v_dict;
-		argv[1].v_type = VAR_UNKNOWN;
-		if (call_callback(item_cb, -1, &rettv, 1, argv) != FAIL)
-		{
-		    if (rettv.v_type == VAR_STRING)
-			itemstr = rettv.vval.v_string;
-		}
-		dict_unref(li->li_tv.vval.v_dict);
-	    }
-	}
-
-	if (itemstr != NULL
-		&& fuzzy_match(itemstr, str, matchseq, &score, matches,
-						MAX_FUZZY_MATCHES, camelcase))
-	{
-	    items[match_count].idx = match_count;
-	    items[match_count].item = li;
-	    items[match_count].score = score;
-
-	    // Copy the list of matching positions in itemstr to a list, if
-	    // 'retmatchpos' is set.
-	    if (retmatchpos)
-	    {
-		int	j = 0;
-		char_u	*p;
-
-		items[match_count].lmatchpos = list_alloc();
-		if (items[match_count].lmatchpos == NULL)
-		    goto done;
-
-		p = str;
-		while (*p != NUL)
-		{
-		    if (!VIM_ISWHITE(PTR2CHAR(p)) || matchseq)
-		    {
-			if (list_append_number(items[match_count].lmatchpos,
-				    matches[j]) == FAIL)
-			    goto done;
-			j++;
-		    }
-		    if (has_mbyte)
-			MB_PTR_ADV(p);
-		    else
-			++p;
-		}
-	    }
-	    ++match_count;
-	}
-	clear_tv(&rettv);
-    }
-
-    if (match_count > 0)
-    {
-	list_T		*retlist;
-
-	// Sort the list by the descending order of the match score
-	qsort((void *)items, (size_t)match_count, sizeof(fuzzyItem_T),
-		fuzzy_match_item_compare);
-
-	// For matchfuzzy(), return a list of matched strings.
-	//	    ['str1', 'str2', 'str3']
-	// For matchfuzzypos(), return a list with three items.
-	// The first item is a list of matched strings. The second item
-	// is a list of lists where each list item is a list of matched
-	// character positions. The third item is a list of matching scores.
-	//	[['str1', 'str2', 'str3'], [[1, 3], [1, 3], [1, 3]]]
-	if (retmatchpos)
-	{
-	    li = list_find(fmatchlist, 0);
-	    if (li == NULL || li->li_tv.vval.v_list == NULL)
-		goto done;
-	    retlist = li->li_tv.vval.v_list;
-	}
-	else
-	    retlist = fmatchlist;
-
-	// Copy the matching strings with a valid score to the return list
-	for (i = 0; i < match_count; i++)
-	{
-	    if (items[i].score == SCORE_NONE)
-		break;
-	    list_append_tv(retlist, &items[i].item->li_tv);
-	}
-
-	// next copy the list of matching positions
-	if (retmatchpos)
-	{
-	    li = list_find(fmatchlist, -2);
-	    if (li == NULL || li->li_tv.vval.v_list == NULL)
-		goto done;
-	    retlist = li->li_tv.vval.v_list;
-
-	    for (i = 0; i < match_count; i++)
-	    {
-		if (items[i].score == SCORE_NONE)
-		    break;
-		if (items[i].lmatchpos != NULL
-		      && list_append_list(retlist, items[i].lmatchpos) == FAIL)
-		    goto done;
-	    }
-
-	    // copy the matching scores
-	    li = list_find(fmatchlist, -1);
-	    if (li == NULL || li->li_tv.vval.v_list == NULL)
-		goto done;
-	    retlist = li->li_tv.vval.v_list;
-	    for (i = 0; i < match_count; i++)
-	    {
-		if (items[i].score == SCORE_NONE)
-		    break;
-		if (list_append_number(retlist, items[i].score) == FAIL)
-		    goto done;
-	    }
-	}
-    }
-
-done:
-    vim_free(items);
-}
-
-/*
- * Do fuzzy matching. Returns the list of matched strings in 'rettv'.
- * If 'retmatchpos' is TRUE, also returns the matching character positions.
- */
-    static void
-do_fuzzymatch(typval_T *argvars, typval_T *rettv, int retmatchpos)
-{
-    callback_T	cb;
-    char_u	*key = NULL;
-    int		ret;
-    int		matchseq = FALSE;
-    long	max_matches = 0;
-    int		camelcase = TRUE;
-
-    if (in_vim9script()
-	    && (check_for_list_arg(argvars, 0) == FAIL
-		|| check_for_string_arg(argvars, 1) == FAIL
-		|| check_for_opt_dict_arg(argvars, 2) == FAIL))
-	return;
-
-    CLEAR_POINTER(&cb);
-
-    // validate and get the arguments
-    if (argvars[0].v_type != VAR_LIST || argvars[0].vval.v_list == NULL)
-    {
-	semsg(_(e_argument_of_str_must_be_list),
-			     retmatchpos ? "matchfuzzypos()" : "matchfuzzy()");
-	return;
-    }
-    if (argvars[1].v_type != VAR_STRING
-	    || argvars[1].vval.v_string == NULL)
-    {
-	semsg(_(e_invalid_argument_str), tv_get_string(&argvars[1]));
-	return;
-    }
-
-    if (argvars[2].v_type != VAR_UNKNOWN)
-    {
-	dict_T		*d;
-	dictitem_T	*di;
-
-	if (check_for_nonnull_dict_arg(argvars, 2) == FAIL)
-	    return;
-
-	// To search a dict, either a callback function or a key can be
-	// specified.
-	d = argvars[2].vval.v_dict;
-	if ((di = dict_find(d, (char_u *)"key", -1)) != NULL)
-	{
-	    if (di->di_tv.v_type != VAR_STRING
-		    || di->di_tv.vval.v_string == NULL
-		    || *di->di_tv.vval.v_string == NUL)
-	    {
-		semsg(_(e_invalid_value_for_argument_str_str), "key",
-						    tv_get_string(&di->di_tv));
-		return;
-	    }
-	    key = tv_get_string(&di->di_tv);
-	}
-	else if ((di = dict_find(d, (char_u *)"text_cb", -1)) != NULL)
-	{
-	    cb = get_callback(&di->di_tv);
-	    if (cb.cb_name == NULL)
-	    {
-		semsg(_(e_invalid_value_for_argument_str), "text_cb");
-		return;
-	    }
-	}
-
-	if ((di = dict_find(d, (char_u *)"limit", -1)) != NULL)
-	{
-	    if (di->di_tv.v_type != VAR_NUMBER)
-	    {
-		semsg(_(e_invalid_value_for_argument_str), "limit");
-		return;
-	    }
-	    max_matches = (long)tv_get_number_chk(&di->di_tv, NULL);
-	}
-
-	if ((di = dict_find(d, (char_u *)"camelcase", -1)) != NULL)
-	{
-	    if (di->di_tv.v_type != VAR_BOOL)
-	    {
-		semsg(_(e_invalid_value_for_argument_str), "camelcase");
-		return;
-	    }
-	    camelcase = tv_get_bool_chk(&di->di_tv, NULL);
-	}
-
-	if (dict_has_key(d, "matchseq"))
-	    matchseq = TRUE;
-    }
-
-    // get the fuzzy matches
-    ret = rettv_list_alloc(rettv);
-    if (ret == FAIL)
-	goto done;
-    if (retmatchpos)
-    {
-	list_T	*l;
-
-	// For matchfuzzypos(), a list with three items are returned. First
-	// item is a list of matching strings, the second item is a list of
-	// lists with matching positions within each string and the third item
-	// is the list of scores of the matches.
-	l = list_alloc();
-	if (l == NULL)
-	    goto done;
-	if (list_append_list(rettv->vval.v_list, l) == FAIL)
-	{
-	    vim_free(l);
-	    goto done;
-	}
-	l = list_alloc();
-	if (l == NULL)
-	    goto done;
-	if (list_append_list(rettv->vval.v_list, l) == FAIL)
-	{
-	    vim_free(l);
-	    goto done;
-	}
-	l = list_alloc();
-	if (l == NULL)
-	    goto done;
-	if (list_append_list(rettv->vval.v_list, l) == FAIL)
-	{
-	    vim_free(l);
-	    goto done;
-	}
-    }
-
-    fuzzy_match_in_list(argvars[0].vval.v_list, tv_get_string(&argvars[1]),
-	    matchseq, key, &cb, retmatchpos, rettv->vval.v_list, max_matches,
-	    camelcase);
-
-done:
-    free_callback(&cb);
-}
-
-/*
- * "matchfuzzy()" function
- */
-    void
-f_matchfuzzy(typval_T *argvars, typval_T *rettv)
-{
-    do_fuzzymatch(argvars, rettv, FALSE);
-}
-
-/*
- * "matchfuzzypos()" function
- */
-    void
-f_matchfuzzypos(typval_T *argvars, typval_T *rettv)
-{
-    do_fuzzymatch(argvars, rettv, TRUE);
-}
-#endif
-
-/*
- * Same as fuzzy_match_item_compare() except for use with a string match
- */
-    static int
-fuzzy_match_str_compare(const void *s1, const void *s2)
-{
-    int		v1 = ((fuzmatch_str_T *)s1)->score;
-    int		v2 = ((fuzmatch_str_T *)s2)->score;
-    int		idx1 = ((fuzmatch_str_T *)s1)->idx;
-    int		idx2 = ((fuzmatch_str_T *)s2)->idx;
-
-    if (v1 == v2)
-	return idx1 == idx2 ? 0 : idx1 > idx2 ? 1 : -1;
-    else
-	return v1 > v2 ? -1 : 1;
-}
-
-/*
- * Sort fuzzy matches by score
- */
-    static void
-fuzzy_match_str_sort(fuzmatch_str_T *fm, int sz)
-{
-    // Sort the list by the descending order of the match score
-    qsort((void *)fm, (size_t)sz, sizeof(fuzmatch_str_T),
-	    fuzzy_match_str_compare);
-}
-
-/*
- * Same as fuzzy_match_item_compare() except for use with a function name
- * string match. <SNR> functions should be sorted to the end.
- */
-    static int
-fuzzy_match_func_compare(const void *s1, const void *s2)
-{
-    int		v1 = ((fuzmatch_str_T *)s1)->score;
-    int		v2 = ((fuzmatch_str_T *)s2)->score;
-    int		idx1 = ((fuzmatch_str_T *)s1)->idx;
-    int		idx2 = ((fuzmatch_str_T *)s2)->idx;
-    char_u	*str1 = ((fuzmatch_str_T *)s1)->str;
-    char_u	*str2 = ((fuzmatch_str_T *)s2)->str;
-
-    if (*str1 != '<' && *str2 == '<')
-	return -1;
-    if (*str1 == '<' && *str2 != '<')
-	return 1;
-    if (v1 == v2)
-	return idx1 == idx2 ? 0 : idx1 > idx2 ? 1 : -1;
-    else
-	return v1 > v2 ? -1 : 1;
-}
-
-/*
- * Sort fuzzy matches of function names by score.
- * <SNR> functions should be sorted to the end.
- */
-    static void
-fuzzy_match_func_sort(fuzmatch_str_T *fm, int sz)
-{
-    // Sort the list by the descending order of the match score
-    qsort((void *)fm, (size_t)sz, sizeof(fuzmatch_str_T),
-		fuzzy_match_func_compare);
-}
-
-/*
- * Fuzzy match 'pat' in 'str'. Returns 0 if there is no match. Otherwise,
- * returns the match score.
- */
-    int
-fuzzy_match_str(char_u *str, char_u *pat)
-{
-    int		score = 0;
-    int_u	matchpos[MAX_FUZZY_MATCHES];
-
-    if (str == NULL || pat == NULL)
-	return 0;
-
-    fuzzy_match(str, pat, TRUE, &score, matchpos,
-				sizeof(matchpos) / sizeof(matchpos[0]), TRUE);
-
-    return score;
-}
-
-/*
- * Fuzzy match the position of string 'pat' in string 'str'.
- * Returns a dynamic array of matching positions. If there is no match,
- * returns NULL.
- */
-    garray_T *
-fuzzy_match_str_with_pos(char_u *str UNUSED, char_u *pat UNUSED)
-{
-#ifdef FEAT_SEARCH_EXTRA
-    int		    score = 0;
-    garray_T	    *match_positions = NULL;
-    int_u	    matches[MAX_FUZZY_MATCHES];
-    int		    j = 0;
-
-    if (str == NULL || pat == NULL)
-	return NULL;
-
-    match_positions = ALLOC_ONE(garray_T);
-    if (match_positions == NULL)
-	return NULL;
-    ga_init2(match_positions, sizeof(int_u), 10);
-
-    if (!fuzzy_match(str, pat, FALSE, &score, matches, MAX_FUZZY_MATCHES, TRUE)
-	    || score == 0)
-    {
-	ga_clear(match_positions);
-	vim_free(match_positions);
-	return NULL;
-    }
-
-    for (char_u *p = pat; *p != NUL; MB_PTR_ADV(p))
-    {
-	if (!VIM_ISWHITE(PTR2CHAR(p)))
-	{
-	    ga_grow(match_positions, 1);
-	    ((int_u *)match_positions->ga_data)[match_positions->ga_len] =
-								    matches[j];
-	    match_positions->ga_len++;
-	    j++;
-	}
-    }
-
-    return match_positions;
-#else
-    return NULL;
-#endif
-}
-
-/*
- * This function splits the line pointed to by `*ptr` into words and performs
- * a fuzzy match for the pattern `pat` on each word. It iterates through the
- * line, moving `*ptr` to the start of each word during the process.
- *
- * If a match is found:
- * - `*ptr` points to the start of the matched word.
- * - `*len` is set to the length of the matched word.
- * - `*score` contains the match score.
- *
- * If no match is found, `*ptr` is updated to the end of the line.
- */
-    int
-fuzzy_match_str_in_line(
-    char_u	**ptr,
-    char_u	*pat,
-    int		*len,
-    pos_T	*current_pos,
-    int		*score)
-{
-    char_u	*str = *ptr;
-    char_u	*strBegin = str;
-    char_u	*end = NULL;
-    char_u	*start = NULL;
-    int		found = FALSE;
-    char	save_end;
-    char_u	*line_end = NULL;
-
-    if (str == NULL || pat == NULL)
-	return found;
-    line_end = find_line_end(str);
-
-    while (str < line_end)
-    {
-	// Skip non-word characters
-	start = find_word_start(str);
-	if (*start == NUL)
-	    break;
-	end = find_word_end(start);
-
-	// Extract the word from start to end
-	save_end = *end;
-	*end = NUL;
-
-	// Perform fuzzy match
-	*score = fuzzy_match_str(start, pat);
-	*end = save_end;
-
-	if (*score > 0)
-	{
-	    *len = (int)(end - start);
-	    found = TRUE;
-	    *ptr = start;
-	    if (current_pos)
-		current_pos->col += (int)(end - strBegin);
-	    break;
-	}
-
-	// Move to the end of the current word for the next iteration
-	str = end;
-	// Ensure we continue searching after the current word
-	while (*str != NUL && !vim_iswordp(str))
-	    MB_PTR_ADV(str);
-    }
-
-    if (!found)
-	*ptr = line_end;
-
-    return found;
-}
-
-/*
- * Search for the next fuzzy match in the specified buffer.
- * This function attempts to find the next occurrence of the given pattern
- * in the buffer, starting from the current position. It handles line wrapping
- * and direction of search.
- *
- * Return TRUE if a match is found, otherwise FALSE.
- */
-    int
-search_for_fuzzy_match(
-    buf_T	*buf,
-    pos_T	*pos,
-    char_u	*pattern,
-    int		dir,
-    pos_T	*start_pos,
-    int		*len,
-    char_u	**ptr,
-    int		*score)
-{
-    pos_T	current_pos = *pos;
-    pos_T	circly_end;
-    int		found_new_match = FALSE;
-    int		looped_around = FALSE;
-    int		whole_line = ctrl_x_mode_whole_line();
-
-    if (buf == curbuf)
-	circly_end = *start_pos;
-    else
-    {
-	circly_end.lnum = buf->b_ml.ml_line_count;
-	circly_end.col = 0;
-	circly_end.coladd = 0;
-    }
-
-    if (whole_line && start_pos->lnum != pos->lnum)
-	current_pos.lnum += dir;
-
-    do
-    {
-
-	// Check if looped around and back to start position
-	if (looped_around && EQUAL_POS(current_pos, circly_end))
-	    break;
-
-	// Ensure current_pos is valid
-	if (current_pos.lnum >= 1 && current_pos.lnum <= buf->b_ml.ml_line_count)
-	{
-	    // Get the current line buffer
-	    *ptr = ml_get_buf(buf, current_pos.lnum, FALSE);
-	    if (!whole_line)
-		*ptr += current_pos.col;
-
-	    // If ptr is end of line is reached, move to next line
-	    // or previous line based on direction
-	    if (*ptr != NULL && **ptr != NUL)
-	    {
-		if (!whole_line)
-		{
-		    // Try to find a fuzzy match in the current line starting
-		    // from current position
-		    found_new_match = fuzzy_match_str_in_line(ptr, pattern,
-						    len, &current_pos, score);
-		    if (found_new_match)
-		    {
-			*pos = current_pos;
-			break;
-		    }
-		    else if (looped_around && current_pos.lnum == circly_end.lnum)
-			break;
-		}
-		else
-		{
-		    if (fuzzy_match_str(*ptr, pattern) > 0)
-		    {
-			found_new_match = TRUE;
-			*pos = current_pos;
-			*len = (int)ml_get_buf_len(buf, current_pos.lnum);
-			break;
-		    }
-		}
-	    }
-	}
-
-	// Move to the next line or previous line based on direction
-	if (dir == FORWARD)
-	{
-	    if (++current_pos.lnum > buf->b_ml.ml_line_count)
-	    {
-		if (p_ws)
-		{
-		    current_pos.lnum = 1;
-		    looped_around = TRUE;
-		}
-		else
-		    break;
-	    }
-	}
-	else
-	{
-	    if (--current_pos.lnum < 1)
-	    {
-		if (p_ws)
-		{
-		    current_pos.lnum = buf->b_ml.ml_line_count;
-		    looped_around = TRUE;
-		}
-		else
-		    break;
-
-	    }
-	}
-	current_pos.col = 0;
-    } while (TRUE);
-
-    return found_new_match;
-}
-
-/*
- * Free an array of fuzzy string matches "fuzmatch[count]".
- */
-    void
-fuzmatch_str_free(fuzmatch_str_T *fuzmatch, int count)
-{
-    int i;
-
-    if (fuzmatch == NULL)
-	return;
-    for (i = 0; i < count; ++i)
-	vim_free(fuzmatch[i].str);
-    vim_free(fuzmatch);
-}
-
-/*
- * Copy a list of fuzzy matches into a string list after sorting the matches by
- * the fuzzy score. Frees the memory allocated for 'fuzmatch'.
- * Returns OK on success and FAIL on memory allocation failure.
- */
-    int
-fuzzymatches_to_strmatches(
-	fuzmatch_str_T	*fuzmatch,
-	char_u		***matches,
-	int		count,
-	int		funcsort)
-{
-    int		i;
-
-    if (count <= 0)
-	return OK;
-
-    *matches = ALLOC_MULT(char_u *, count);
-    if (*matches == NULL)
-    {
-	fuzmatch_str_free(fuzmatch, count);
-	return FAIL;
-    }
-
-    // Sort the list by the descending order of the match score
-    if (funcsort)
-	fuzzy_match_func_sort((void *)fuzmatch, (size_t)count);
-    else
-	fuzzy_match_str_sort((void *)fuzmatch, (size_t)count);
-
-    for (i = 0; i < count; i++)
-	(*matches)[i] = fuzmatch[i].str;
-    vim_free(fuzmatch);
-
-    return OK;
-}
